[Support]: clean up

+++ lib/Support/Hash.hpp
@@ -16,7 +16,7 @@ template <typename T>
   requires requires { std::hash<T>(); }
 struct Hash<T> : std::hash<T> {};
 
-template <typename T> auto hash_combine(usize hash, const T &value) -> usize {
+template <typename T> auto hash_combine(u64 hash, const T &value) -> u64 {
   boost::hash_combine(hash, Hash<T>()(value));
   return hash;
 }
@@ -26,8 +26,8 @@ template <typename T> auto hash_combine(usize hash, const T &value) -> usize {
 
 #define REN_DEFINE_TYPE_HASH(Type, ...)                                        \
   template <> struct Hash<Type> {                                              \
-    auto operator()(const Type &value) const noexcept -> usize {               \
-      usize seed = 0;                                                          \
+    auto operator()(const Type &value) const noexcept -> u64 {                 \
+      u64 seed = 0;                                                            \
       BOOST_PP_SEQ_FOR_EACH(REN_HASH_COMBINE_FIELD, value,                     \
                             BOOST_PP_VARIADIC_TO_SEQ(__VA_ARGS__))             \
       return seed;                                                             \
